+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Support for more information in order notes
+* New `product.quote` and `product.quote.confirmation` events added
 
 ### Changed